Revert "migrate to AndroidX fix Navigation drawer inflation error"
This reverts commit b5905c7f4a941822089e8fd8383527fbb385b77f.
