+++ app/src/main/AndroidManifest.xml
@@ -7,17 +7,16 @@
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
 
     <application
+        android:requestLegacyExternalStorage="true"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
-        android:largeHeap="true"
-        android:requestLegacyExternalStorage="true"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
+        android:largeHeap="true"
         android:theme="@style/ActionBarTheme.NoActionBar">
-
-        <activity android:name=".background.AlphabetBaseShow" />
-        <activity android:name=".infront.CharacterSelection" />
+        <activity android:name=".background.alphabetBaseShow" />
+        <activity android:name=".infront.CharacterSelection"/>
         <activity android:name=".infront.SplashScreen">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -28,10 +27,9 @@
         <activity
             android:name=".infront.LanguageSelection"
             android:configChanges="orientation|screenSize" />
-        <activity
-            android:name=".infront.AlphabetShow"
+        <activity android:name=".infront.alphabetshow"
 
-            android:configChanges="orientation|screenSize" />
+            android:configChanges="orientation|screenSize"/>
     </application>
 
 </manifest>
