+++ app/src/main/java/com/example/myapplication/background/backgrounds.java
@@ -3,7 +3,7 @@ package com.example.myapplication.background;
 import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
-import androidx.appcompat.app.AppCompatActivity;
+import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.widget.Toast;
