+++ app/src/main/java/com/example/myapplication/infront/CharacterSelection.java
@@ -5,7 +5,7 @@ import android.content.Intent;
 import android.graphics.Point;
 import android.os.Build;
 import android.os.Bundle;
-import androidx.annotation.RequiresApi;
+import android.support.annotation.RequiresApi;
 import android.view.Display;
 import android.view.View;
 import android.view.animation.Animation;
@@ -14,6 +14,7 @@ import android.widget.LinearLayout;
 
 import com.example.myapplication.R;
 import com.example.myapplication.background.Animator;
+import com.example.myapplication.background.alphabetBaseShow;
 
 import java.util.Random;
 
@@ -68,7 +69,7 @@ public class CharacterSelection extends Activity {
                @Override
                public void onClick(View v) {
                    Intent intent;
-                   intent = new Intent(CharacterSelection.this, AlphabetShow.class);
+                   intent = new Intent(CharacterSelection.this, alphabetshow.class);
                    intent.putExtra(getResources().getString(R.string.practice_string),((Button)v).getText());
                    startActivity(intent);
