+++ app/src/main/java/com/example/myapplication/infront/LanguageSelection.java
@@ -4,12 +4,14 @@ import android.content.Intent;
 import android.content.pm.ActivityInfo;
 import android.content.res.Configuration;
 import android.os.Bundle;
-import androidx.appcompat.app.AppCompatActivity;
+import android.support.v7.app.AppCompatActivity;
 import android.view.Menu;
 import android.view.View;
 import android.widget.Button;
 
+import com.example.myapplication.infront.CharacterSelection;
 import com.example.myapplication.R;
+import com.example.myapplication.infront.SplashScreen;
 
 import java.util.Locale;
