+++ app/src/main/java/com/example/myapplication/infront/SplashScreen.java
@@ -9,14 +9,15 @@ import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.speech.tts.TextToSpeech;
-import androidx.annotation.Nullable;
-import androidx.core.app.ActivityCompat;
-import androidx.core.content.ContextCompat;
+import android.support.annotation.Nullable;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.content.ContextCompat;
 import android.util.DisplayMetrics;
 import android.widget.Toast;
 
 import com.example.myapplication.R;
 import com.example.myapplication.background.ScoreDbHelper;
+import com.example.myapplication.background.alphabetBaseShow;
 
 public class SplashScreen extends Activity implements TextToSpeech.OnInitListener {
