+++ app/src/main/res/layout/activity_alphabet_show.xml
@@ -1,19 +1,29 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.drawerlayout.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/drawerLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:fitsSystemWindows="true"
-    tools:context=".background.AlphabetBaseShow"
-    tools:openDrawer="start">
+    tools:context=".background.alphabetBaseShow">
+
+    <android.support.design.widget.NavigationView
+        android:id="@+id/nav_view"
+        android:layout_width="wrap_content"
+        android:layout_height="match_parent"
+        android:layout_gravity="start"
+        android:layout_marginTop="55dp"
+        app:headerLayout="@layout/header"
+        app:menu="@menu/colors"
+        android:clickable="true">
+
+    </android.support.design.widget.NavigationView>
 
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <androidx.appcompat.widget.Toolbar
+        <android.support.v7.widget.Toolbar
             android:id="@+id/PracticeToolbar"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -81,16 +91,5 @@
             android:onClick="practiceOnClick"
             app:button_color="@color/ButtonBg"
             app:image="@drawable/ic_reset" />
-
     </RelativeLayout>
-
-    <com.google.android.material.navigation.NavigationView
-        android:id="@+id/nav_view"
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentBottom="true"
-        android:layout_gravity="start"
-        app:headerLayout="@layout/header"
-        app:menu="@menu/colors" />
-</androidx.drawerlayout.widget.DrawerLayout>
+</android.support.v4.widget.DrawerLayout>
