+++ app/src/main/res/layout/activity_character_selection.xml
@@ -7,7 +7,7 @@
     android:orientation="vertical"
     tools:context=".infront.CharacterSelection">
 
-    <androidx.appcompat.widget.Toolbar
+    <android.support.v7.widget.Toolbar
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="@color/ActionbarBg"
