+++ app/src/main/res/layout/activity_first.xml
@@ -7,7 +7,7 @@
     tools:context=".infront.LanguageSelection"
     android:orientation="vertical">
 
-<androidx.appcompat.widget.Toolbar
+<android.support.v7.widget.Toolbar
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     app:title="Language Selection"
