+++ app/src/main/res/layout/header.xml
@@ -1,15 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="176dp"
-    android:orientation="vertical">
+    android:orientation="vertical" android:layout_width="match_parent"
+    android:layout_height="match_parent">
 
     <TextView
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="128dp"
         android:background="@drawable/yellow"
         android:text="SELECT COLOR"
         android:textColor="@color/Black"
-        android:textSize="40sp"
+        android:textSize="40dp"
         android:textStyle="bold" />
 </LinearLayout>
