+++ app/src/main/res/menu/colors.xml
@@ -16,7 +16,8 @@
         <item
             android:id="@+id/red"
             android:icon="@drawable/ic_save"
-            android:title="RED"/>
+            android:title="RED"
+            android:onClick="toast"/>
 
     </group>
 </menu>
