+++ app/src/main/res/menu/menu.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".background.AlphabetBaseShow">
+    tools:context=".background.alphabetBaseShow">
 
     <item
         android:id="@+id/action_previous"
