+++ app/src/main/res/values/styles.xml
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <style name="ActionBarTheme" parent="Theme.AppCompat.Light.DarkActionBar">
+    <style name="ActionBarTheme" parent="Theme.AppCompat.Light.NoActionBar">
         <item name="android:actionBarStyle">@style/MyActionBar</item>
+        <item name="android:windowActionBar">false</item>
 
         <!-- Support library compatibility -->
         <item name="actionBarStyle">@style/MyActionBar</item>
@@ -24,7 +25,7 @@
         <item name="android:textColor">@color/White</item>
     </style>
 
-    <style name="ActionBarTheme.NoActionBar" parent="ActionBarTheme">
+    <style name="ActionBarTheme.NoActionBar">
         <item name="windowActionBar">false</item>
         <item name="windowNoTitle">true</item>
     </style>
